reallocate DO pins 2,8,10 for IR circuits. HC05 power switch.
Pin 10 is the input from the IR sensor, LOW = no ball.
Pin 2 is input for testing the ball circuits.
Pin 8 turns on power to the HC-05. This allows for downloads by not turning on the HC-05 after the reset for code loading. It is turned on by our setup code.